+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Add check for `change_column_default`
 - Add check for `add_unique_key` (Active Record >= 7.1)
 - Add check for `add_column` with stored generated columns